feat: 커뮤니티 게시판 카드용 시간 변환 함수 (#44)
